Functions and operators utilities
Adds a module containing useful functions to create
or change functions, to use in HOF for example.

operators, composition functions, arity changing functions.

Adds a `toArray` method to tuples and arrays to ease
polymorphism:

* for tuples, returns a copy of the underlying data
* for arrays, returns the array itself

Adds tooling method to FunctionReference:
an `arity` function, and a function that apply arguments, binding them or
invoking the function according to the number.

Add a toArray method and a destruct method to several unions, structs
and arrays.