Adds default values for fields in struct and union
This PR allows to define default values for fields in struct and union.

It define a new IR element, `Member`, to encapsulate the name and default
value of the field, as well as a new abstract node `TypeWithMembers` to
define common behavior for struct and union values regarding their
fields.

Since we have a AST object, it is now possible to document fields:

    struct Foo = {
      ----
      the foo
      ----
      foo,

      ----
      the bar
      ----
      bar
    }

These comments are displayed in the golodoc.

The main feature is implemented by creating alternative factory methods
in the module declaring the struct or the union. See
<https://github.com/eclipse/golo-lang/issues/419> for some details.
For instance, a struct defined as:

    struct Foo = {a, b=42, c=f(a)}

will create several factories:

    function Foo = |a, b, c| -> module.types.Foo(a, b, c)
    function Foo = |a| -> module.types.Foo(a, 42, f(a))
    function Foo = -> raise("Can't call the default constructor...")

while

    struct Bar = {a, b=42}

will generate:

    function Bar = -> module.types.Bar(null, 42)
    function Bar = |a| -> module.types.Bar(a, 42)
    function Bar = |a, b| -> module.types.Bar(a, b)