submodules
A submodule is an inner class of its defining module.
Function from the submodule can access the local function of the parent
module, and the functions in the parent module can access the local
functions in the submodule.

refactor bytecode generation for submodules
Implicit import of submodules and parent module